refactor : GroupingCard를 components 폴더로 이동
