+++ src/components/Card/GroupingCard.client.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { formatDate } from '../util';
+import { formatDate } from '../../app/grouping/util';
 import { Spacing } from '@/components/common/Spacing';
 import { Flex } from '@/components/Layout';
 import clsx from 'clsx';
